Use follow_imports=silent in pre-commit mypy
This brings this file in line with the default configuration (mypy.ini)
and allows better typing. Without this change, the next commit could not
pass both the CI check (which uses `mypy.ini`)  and the pre-commit check
(which uses `.pre-commit.mypy.ini`) because `mypy.ini` would require a
type:ignore comment but `.pre-commit.mypy.ini` would reject it as
unnecessary.